backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
langchain-core
langchain-google-genai
langgraph
python-dotenv
pydantic[email]
